Bugfix for #452 , error opening config on KDE.
This bug happens when trying to launch Fcitx Configuration
in KDE, with the kcm-fcitx or kde-config-fcitx module already
installed. It opens a dialog that says this:
```
You're currently running KDE, but KCModule for fcitx couldn't be found, the package name of this KCModule is usually kcm-fcitx or kde-config-fcitx.
```

It is caused by checking for the KDE version by using the
`xprop -root KDE_SESSION_VERSION` command on a system where
that xprop is not defined. This fix has the fcitx-configtool
script use the `KDE_SESSION_VERSION` environment variable
by default, and then fail back to `xprop` if it doesn't work.

This fix also introduces some more precise error handling for
the KDE/KCM detection logic, including error messages that
should be less likely to be misleading. (The new error messages
are English-only (untranslated) in this commit.)